refactor: split types.ts into domain modules + decompose SprintConfig
- Split monolithic types.ts (289 lines) into 12 domain modules in src/types/:
  schemas, planning, execution, quality, escalation, metrics, drift, git,
  documentation, ceremonies, config, index
- Decompose SprintConfig into sub-config interfaces:
  GitConfig, SessionConfig, ExecutionLimits, QualityGateSettings
  SprintConfig extends all three for backward compatibility
- Original types.ts becomes a re-export shim — zero import changes needed
- 472 tests pass, types clean
