+++ src/types/drift.ts
@@ -1,0 +1,11 @@
+export interface DriftReport {
+  totalFilesChanged: number;
+  plannedChanges: number;
+  unplannedChanges: string[];
+  driftPercentage: number;
+}
+
+export interface DriftIncident {
+  issueNumber: number;
+  files: string[];
+}
