+++ src/types/git.ts
@@ -1,0 +1,11 @@
+export interface Worktree {
+  path: string;
+  branch: string;
+  issueNumber: number;
+}
+
+export interface DiffStat {
+  linesChanged: number;
+  filesChanged: number;
+  files: string[];
+}
