+++ src/types/metrics.ts
@@ -1,0 +1,11 @@
+export interface SprintMetrics {
+  planned: number;
+  completed: number;
+  failed: number;
+  pointsPlanned: number;
+  pointsCompleted: number;
+  velocity: number;
+  avgDuration: number;
+  firstPassRate: number;
+  driftIncidents: number;
+}
