+++ src/types/quality.ts
@@ -1,0 +1,11 @@
+export interface QualityCheck {
+  name: string;
+  passed: boolean;
+  detail: string;
+  category: "lint" | "test" | "type" | "build" | "diff" | "other";
+}
+
+export interface QualityResult {
+  passed: boolean;
+  checks: QualityCheck[];
+}
